remove union.js
